AI_Agent/requirements.txt
langchain==1.0.5
langchain-community
langchain-groq
python-dotenv
pydantic
duckduckgo-search
wikipedia
